feat: shortcut 기능

+++ backend/src/main/java/com/ssafy/teongbin/shortcut/repository/ShortcutRepository.java
@@ -4,8 +4,7 @@ import com.ssafy.teongbin.shortcut.entity.Shortcut;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface ShortcutRepository extends JpaRepository<Shortcut, Long> {
-    public List<Shortcut> findByUserId(Integer UserId);
+    public List<Shortcut> findByUserId(Long UserId);
 }
